pyproject.toml
[project]
name = "trivme-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14.2"
dependencies = [
    "google-genai>=1.59.0",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "pydantic>=2.12.5",
]
